Improved efficiency of preview
The currrent process: 1. Maintain a copy of the font as a TTFont object in memory. This is separate from the working copy. 2. When a preview is needed, pass this to Xgridfit, which will compile instructions for the single glyph that needs to be previewed. 3.) Xgridfit inserts the compiled instructions in the font. 4.) Xgridfit calls fontTools to subset the font and 5.) calls fontTools to write the subsetted font to a SpooledTemporaryFile. 6.) The handle to this file is returned to ygt, which 7.) passes it to ygPreview for display. For Junicode (4000+ glyphs), this happens in about 0.66 seconds.